Moving back to old file and using subprocess to run test isolated
